backend/requirements.txt
fastapi==0.95.2
requests==2.28.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests>=2.28.0
python-dotenv>=1.0.0